refactor(common): restructure serialization
