+++ common/src/main/kotlin/de/hpi/cloud/common/types/Image.kt
@@ -1,40 +1,17 @@
 package de.hpi.cloud.common.types
 
-import de.hpi.cloud.common.Context
 import de.hpi.cloud.common.Persistable
-import de.hpi.cloud.common.protobuf.build
-import de.hpi.cloud.common.serializers.EnumSerializer
-import de.hpi.cloud.common.serializers.UriSerializer
-import de.hpi.cloud.common.utils.parseUri
+import de.hpi.cloud.common.serializers.json.EnumSerializer
+import de.hpi.cloud.common.serializers.json.UrlSerializer
 import kotlinx.serialization.Serializable
-import java.net.URI
-import de.hpi.cloud.common.v1test.Image as ProtoImage
+import java.net.URL
 
 @Serializable
 data class Image(
-    val source: Map<Size, @Serializable(with = UriSerializer::class) URI>,
+    val source: Map<Size, @Serializable(with = UrlSerializer::class) URL>,
     val alt: L10n<String>,
     val aspectRatio: Float? = null
 ) : Persistable<Image>() {
-    object ProtoSerializer : Persistable.ProtoSerializer<Image, ProtoImage> {
-        override fun fromProto(proto: ProtoImage, context: Context): Image {
-            return Image(
-                source = mapOf(Size.ORIGINAL to proto.source.parseUri()),
-                alt = L10n.single(context, proto.alt),
-                aspectRatio = proto.aspectRatio
-            )
-        }
-
-        override fun toProto(persistable: Image, context: Context): ProtoImage =
-            toProto(persistable, context, Size.ORIGINAL)
-
-        fun toProto(persistable: Image, context: Context, size: Size = Size.ORIGINAL): ProtoImage =
-            ProtoImage.newBuilder().build(persistable) {
-                source = it.source.bestMatch(size).toString()
-                alt = it.alt[context]
-                it.aspectRatio?.let { a -> aspectRatio = a }
-            }
-    }
 
     @Serializable(with = Size.Serializer::class)
     enum class Size {
@@ -42,15 +19,4 @@ data class Image(
 
         object Serializer : EnumSerializer<Size>(Size::class, ORIGINAL)
     }
-}
-
-fun Image.toProto(context: Context): ProtoImage = Image.ProtoSerializer.toProto(this, context)
-
-
-fun <T> Map<Image.Size, T>.bestMatch(size: Image.Size): T {
-    require(isNotEmpty()) { "At least one size must be available" }
-
-    return this[size]
-        ?: this[Image.Size.ORIGINAL]
-        ?: values.first()
 }
