+++ common/src/main/kotlin/de/hpi/cloud/common/types/MarkupContent.kt
@@ -1,11 +1,8 @@
 package de.hpi.cloud.common.types
 
-import de.hpi.cloud.common.Context
 import de.hpi.cloud.common.Persistable
-import de.hpi.cloud.common.protobuf.build
-import de.hpi.cloud.common.serializers.EnumSerializer
+import de.hpi.cloud.common.serializers.json.EnumSerializer
 import kotlinx.serialization.Serializable
-import de.hpi.cloud.common.v1test.MarkupContent as ProtoMarkupContent
 
 @Serializable
 data class MarkupContent(
@@ -18,20 +15,6 @@ data class MarkupContent(
         fun markdown(content: String): MarkupContent = MarkupContent(Type.MARKDOWN, content)
     }
 
-    object ProtoSerializer : Persistable.ProtoSerializer<MarkupContent, ProtoMarkupContent> {
-        override fun fromProto(proto: ProtoMarkupContent, context: Context): MarkupContent {
-            return MarkupContent(
-                type = Type.fromMime(proto.type),
-                content = proto.content
-            )
-        }
-
-        override fun toProto(persistable: MarkupContent, context: Context): ProtoMarkupContent =
-            ProtoMarkupContent.newBuilder().build(persistable) {
-                type = it.type.mime
-                content = it.content
-            }
-    }
 
     @Serializable(with = Type.Serializer::class)
     enum class Type(val mime: String) {
@@ -50,5 +33,3 @@ data class MarkupContent(
         object Serializer : EnumSerializer<Type>(Type::class, PLAIN, { it.mime })
     }
 }
-
-fun MarkupContent.toProto(context: Context): ProtoMarkupContent = MarkupContent.ProtoSerializer.toProto(this, context)
