+++ common/src/main/kotlin/de/hpi/cloud/common/utils/Uri.kt
@@ -2,11 +2,21 @@ package de.hpi.cloud.common.utils
 
 import java.net.URI
 import java.net.URISyntaxException
+import java.net.URL
 
 fun String.parseUri(): URI = URI(this)
 fun String.tryParseUri(): URI? {
     return try {
         parseUri()
+    } catch (e: URISyntaxException) {
+        null
+    }
+}
+
+fun String.parseUrl(): URL = URL(this)
+fun String.tryParseUrl(): URL? {
+    return try {
+        parseUrl()
     } catch (e: URISyntaxException) {
         null
     }
